requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-speech==1.34.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy>=1.26.0
msrest==0.7.1